requirements.txt
accelerate==0.27.2
einops==0.7.0
einops-exts==0.0.4
iopath==0.1.10
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.0
peft==0.4.0
pillow==10.2.0
protobuf==4.25.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.1.99
tiktoken==0.6.0
timm==0.9.12
tokenizers==0.15.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.36.2
